==> ensemblemodelsanddeployment/1.Preprocessing.ipynb <==
import os 
import json
import numpy as np
import pandas as pd
from sklearn.externals import joblib
from sklearn.model_selection import train_test_split, GridSearchCV
from sklearn.base import BaseEstimator, TransformerMixin
from sklearn.ensemble import RandomForestClassifier

from sklearn.pipeline import make_pipeline

import warnings
warnings.filterwarnings("ignore")

from sklearn.feature_selection import VarianceThreshold
import seaborn as sns
import matplotlib.pyplot as plt
from sklearn.ensemble import RandomForestRegressor, RandomForestClassifier
from sklearn.preprocessing import LabelBinarizer
from collections import defaultdict
import pandas as pd
from mlxtend.feature_selection import SequentialFeatureSelector as SFS
from sklearn.metrics import roc_auc_score, mean_squared_error

import os
# Office
os.chdir('/home/exe002/Desktop/Prudhviraju/Code/LoanPrediction')
data = pd.read_csv('train_u6lujuX_CVtuZ9i.csv')

## Listing the columns persent in the dataset

data=data.reset_index(drop=True)

### Handling Null Values

data.loc[data['Gender'].isnull(),'Gender']='Trasgender'
data.loc[data['Dependents'].isnull(),'Dependents']='0'
data.loc[data['Education'].isnull(),'Education']='No Education'
data.loc[data['Married'].isnull(),'Married']='No'
data.loc[data['Self_Employed'].isnull(),'Self_Employed']='No'
data=data.drop(index=data.loc[data['LoanAmount'].isnull(),:].index)
data.loc[data['Loan_Amount_Term'].isnull(),'Loan_Amount_Term']=0
data.loc[data['Credit_History'].isnull(),'Credit_History']=0

Numeric_columns=data.select_dtypes(include=['int64','float64'])
categorical_columns=data.select_dtypes(include=['object'])
print(Numeric_columns.columns)
print(categorical_columns.columns)

## Constant Values in Numeric Columns

from sklearn.feature_selection import VarianceThreshold

#Using a Built In Function
sel = VarianceThreshold(threshold=0)
sel.fit(Numeric_columns)  # fit finds the features with zero variance
print('Total # of non constant features',sum(sel.get_support()))
sel.transform(Numeric_columns)
Nonconstant_Numeric=Numeric_columns.loc[:,list(sel.get_support())]

## Constant Values in Categorical Columns¶

unique_columns=[]
for i in categorical_columns.columns:
    if len(categorical_columns[i].unique())==1:
        unique_columns.append(i)
print(unique_columns)

## Merging the two dataframes

new_df=pd.concat([categorical_columns,Nonconstant_Numeric],axis=1)
new_df.isnull().sum()

## Removing the unique key columns

unique_columns=[i for i in new_df.columns if len(new_df[i].unique())==new_df.shape[0]]
print(unique_columns)

## Dropping the unique key columns

new_df=new_df.drop(unique_columns,axis=1)

## Type conversion

new_df['Credit_History']=new_df['Credit_History'].astype(str)

## Univariate Analysis

IQR=new_df['ApplicantIncome'].quantile(0.75)-new_df['ApplicantIncome'].quantile(0.25)
upper_fence=new_df['ApplicantIncome'].quantile(0.75)+(IQR*3)
lower_fence=new_df['ApplicantIncome'].quantile(0.75)-(IQR*3)
print(IQR,'  ', upper_fence,'  ',lower_fence)
new_df=new_df.loc[(new_df['ApplicantIncome']<=upper_fence) & (new_df['ApplicantIncome']>=lower_fence), :]

from sklearn.preprocessing import MinMaxScaler
from collections import defaultdict
encoding_minmax=defaultdict(MinMaxScaler)
features_num=['ApplicantIncome']
for feat in features_num:
    encoded_df=pd.DataFrame()
    #d=pd.DataFrame(encoding_minmax(feat).fit_transform(new_df[feat]))
    d=pd.DataFrame(encoding_minmax[feat].fit_transform(new_df[[feat]]))
    test_column=d.columns.values
    list_column=[feat+'_'+'minmax' for j in test_column]
    d.columns=list_column
    encoded_df=pd.concat([encoded_df,d],axis=1).reset_index(drop=True)
    new_df.drop(feat,axis=1,inplace=True)
    new_df=new_df.reset_index(drop=True)
    new_df=pd.concat([new_df,encoded_df],axis=1)
import pickle
with open('minmax_pickle.pkl','wb') as f:
    pickle.dump(encoding_minmax,f)

encoding_minmax

from sklearn.preprocessing import StandardScaler
from collections import defaultdict
encoding_standardscaler=defaultdict(StandardScaler)
features_num=['LoanAmount']
for feat in features_num:
    encoded_df=pd.DataFrame()
    #d=pd.DataFrame(encoding_minmax(feat).fit_transform(new_df[feat]))
    d=pd.DataFrame(encoding_standardscaler[feat].fit_transform(new_df[[feat]]))
    test_column=d.columns.values
    list_column=[feat+'_'+'minmax' for j in test_column]
    d.columns=list_column
    encoded_df=pd.concat([encoded_df,d],axis=1).reset_index(drop=True)
    new_df.drop(feat,axis=1,inplace=True)
    new_df=new_df.reset_index(drop=True)
    new_df=pd.concat([new_df,encoded_df],axis=1)

with open('standardscaler_pickle.pkl','wb') as f:
    pickle.dump(encoding_standardscaler,f)

## One Hot Encoding

from sklearn.preprocessing import LabelBinarizer
from collections import defaultdict
import pandas as pd
encoding_binarizer=defaultdict(LabelBinarizer)
col_nominal=['Self_Employed','Married','Gender']
for feat in col_nominal:
    encoded_df=pd.DataFrame()
    d=pd.DataFrame(encoding_binarizer[feat].fit_transform(new_df[feat]))
    test_column=d.columns.values
    list_columns=[feat+'_'+str(j) for j in test_column]
    #print(list_columns)
    d.columns=list_columns
    #print(d.head())
    encoded_df=pd.concat([encoded_df,d],axis=1).reset_index(drop=True)
    new_df.drop(feat,axis=1,inplace=True)
    new_df=new_df.reset_index(drop=True)
    new_df=pd.concat([new_df,encoded_df],axis=1)

with open('binarizer_pickle.pkl','wb') as f:
    pickle.dump(encoding_binarizer,f)

encoding_binarizer

 ### Label Encoding

from sklearn.preprocessing import LabelEncoder
from collections import defaultdict
encoding_labelencoder=defaultdict(LabelEncoder)
col_ordinal=['Education','Property_Area','Loan_Status','Dependents','Credit_History']
for feat in col_ordinal:
    encoded_df=pd.DataFrame()
    d=pd.DataFrame(encoding_labelencoder[feat].fit_transform(new_df[feat]))
    test_column=d.columns.values
    list_columns=[feat+'_'+str(j) for j in test_column]
    d.columns=list_columns
    encoded_df=pd.concat([encoded_df,d],axis=1).reset_index(drop=True)
    new_df.drop(feat,axis=1,inplace=True)
    new_df=new_df.reset_index(drop=True)
    new_df=pd.concat([new_df,encoded_df],axis=1)
#with open('C:\\Users\\prudi\\Desktop\\Data Sets\\Loan Prediction\\labelencoder_pickle.pkl','wb') as f:
#    pickle.dump(encoding_labelencoder,f)
with open('labelencoder_pickle.pkl','wb') as f:
    pickle.dump(encoding_labelencoder,f)

new_df.columns
